+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- **feat: bike plan implementation** - create bike plan training plan
 - **feat: general-scheduler** - migrate scheduler from run plan to general function for each dicipline to utilize
 - **fix: run plan refactor** - removed schedule.ts and moved to generators.ts, updated db table
 - **feat: run-plan-generator** - create run plan off calculation